requirements.txt
streamlit
pandas
matplotlib
openpyxl
openai
python-dotenv
pytest